requirements.txt
fastapi
uvicorn
streamlit
pandas
scikit-learn
pydantic
sqlalchemy
prophet
twilio
spacy
python-dotenv
python-multipart
